Decompose app_frame/handlers/sideboard_guide.py into 3 handler mixins
Split the 585-LOC SideboardGuideHandlers mixin along its three natural
seams into one-mixin-per-file modules matching the handlers/ convention:

- sideboard_guide_persistence.py (SideboardGuidePersistenceHandlers):
  outboard/guide state I/O, pin indicator, and pin action.
- sideboard_guide_entries.py (SideboardGuideEntryHandlers):
  dialog-driven entry CRUD, exclusions, flex slots, and view refresh.
- sideboard_guide_io.py (SideboardGuideImportExportHandlers):
  CSV import/export, matchup regex parsing, and threaded workers.

All three are composed into AppFrame and registered in handlers/__init__.py.
No behavior change: methods remain on the composed AppFrame, so cross-cluster
calls (e.g. persistence migration using _parse_card_text) resolve via self.
No new protocol; AppFrameProto already types the host.
